+++ src/internal/internalAPIs.ts
@@ -1,3 +1,5 @@
+import { HostClientType } from '../public/constants';
+import { ErrorCode, SdkError } from '../public/interfaces';
 import { userOriginUrlValidationRegExp, defaultSDKVersionForCompatCheck } from './constants';
 import { GlobalVars } from './globalVars';
 import { generateRegExpFromUrls, compareSDKVersions } from './utils';
@@ -33,6 +35,32 @@ export function isAPISupportedByPlatform(requiredVersion: string = defaultSDKVer
     return false;
   }
   return value >= 0;
+}
+
+/**
+ * Helper function to identify if host client is either android or ios
+ */
+ export function isHostClientMobile(): boolean {
+  if (GlobalVars.hostClientType == HostClientType.android || GlobalVars.hostClientType == HostClientType.ios) {
+    return true;
+  }
+  return false;
+}
+
+/**
+ * Helper function which indicates if current API is supported on mobile or not.
+ * @returns SdkError if host client is not android/ios or if the requiredVersion is not
+ *          supported by platform or not. Null is returned in case of success.
+ */
+export function isApiSupportedOnMobile(requiredVersion: string = defaultSDKVersionForCompatCheck): SdkError {
+  if (!isHostClientMobile()) {
+    const notSupportedError: SdkError = { errorCode: ErrorCode.NOT_SUPPORTED_ON_PLATFORM };
+    return notSupportedError;
+  } else if (!isAPISupportedByPlatform(requiredVersion)) {
+    const oldPlatformError: SdkError = { errorCode: ErrorCode.OLD_PLATFORM };
+    return oldPlatformError;
+  }
+  return null;
 }
 
 /**
